v0.10

+++ 3VV_Vika3_Pizza_3-Layer/include/models/Extra.h
@@ -14,22 +14,16 @@ class Extra
 
         int getIdNumber() const;
         string getName() const;
-        double getPriceCategory() const;
+        int getPriceCategory() const;
         bool getActiveState() const;
-        //int getLines() const;
 
         void setIdNumber(int newIdNumber);
         void setName(string newName);
-        //void setPrice(double newPrice);
         void setPriceCategory(int newPriceCategory);
         void setActiveState(bool newState);
 
-        //Extra* readFile();
-
         friend istream&  operator >> (istream& in, Extra& extra);
         friend ostream&  operator << (ostream& out, const Extra& extra);
-        //friend ifstream& operator >> (ifstream& fin, Extra& extra);
-        //friend ofstream& operator << (ofstream& fout, const Extra& extra);
 
     private:
         static const int MAX_STRING_LENGTH = 24;
